nuss_lee_project - Backup/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
aiofiles
Pillow
google-api-python-client  